Standardized command structure and organized terminal commands
