+++ src/components/shared/Terminal.tsx
@@ -2,7 +2,8 @@ import _ from 'lodash';
 import React, { useState } from 'react';
 import '../../styles/Terminal.scss';
 import '../../styles/global.scss';
-import { Directory, File, FileSystemObject } from './globalTypes';
+import { Directory } from './globalTypes';
+import { executeCommand, TerminalCommandResult } from './TerminalCommands';
 
 function Terminal(prop: {
   fileSystem: Directory;
@@ -15,331 +16,7 @@ function Terminal(prop: {
   const [inputHistory, setInputHistory] = useState<string[]>([]);
   const [inputHistoryIndex, setInputHistoryIndex] = useState(0);
 
-  function executeList(flags: string, path: string): [string[], string[]] {
-    const displayAll = flags.includes('a');
-    const longFormat = flags.includes('l');
-    const fsObject = path.startsWith('/')
-      ? prop.fileSystem.getFileSystemObjectFromPath(path)
-      : prop.currentWorkingDirectory.getFileSystemObjectFromPath(path);
-
-    if (!fsObject) {
-      return [[], [`ls: '${path}': No such file or directory`]];
-    }
-    if (!fsObject.isDirectory) {
-      return [[fsObject.name], []];
-    }
-    return [
-      (fsObject as Directory).getChildrenNames(displayAll, longFormat),
-      [],
-    ];
-  }
-
-  function executeCd(path: string): string[] {
-    // Make a copy to avoid mutating the original (it's a state variable)
-    const fileSystemCopy = _.cloneDeep(prop.fileSystem);
-    const cwdCopy = fileSystemCopy.getFileSystemObjectFromPath(
-      prop.currentWorkingDirectory.path
-    ) as Directory;
-    const newCwd = fileSystemCopy.changeCurrentWorkingDirectory(cwdCopy, path);
-
-    if (newCwd instanceof Directory) {
-      prop.setFileSystem(fileSystemCopy);
-      prop.setCurrentWorkingDirectory(newCwd);
-    } else {
-      return [newCwd];
-    }
-    return [];
-  }
-
-  function executeCat(path: string): [string[], string[]] {
-    const file = path.startsWith('/')
-      ? prop.fileSystem.getFileSystemObjectFromPath(path)
-      : prop.currentWorkingDirectory.getFileSystemObjectFromPath(path);
-
-    if (!file) {
-      return [[], [`cat: '${path}': No such file or directory`]];
-    }
-    if (file.isDirectory) {
-      return [[], [`cat: '${path}': Is a directory`]];
-    }
-    // console.log(((file as File).content ||= '\n'));
-    return [[((file as File).content ||= '\n')], []];
-  }
-
-  function executeTouch(path: string, directory = false): string[] {
-    // Make a copy to avoid mutating the original (it's a state variable)
-    const fileSystemCopy = _.cloneDeep(prop.fileSystem);
-    const cwdCopy = fileSystemCopy.getFileSystemObjectFromPath(
-      prop.currentWorkingDirectory.path
-    ) as Directory;
-    let newFileName = '';
-    let dir;
-
-    // If the path is absolute, find the directory to create the file in
-    // Otherwise, use the current working directory
-    if (path.includes('/')) {
-      const pathArr = path.split('/');
-      newFileName = pathArr.pop() || '';
-      if (path.startsWith('/')) {
-        dir = fileSystemCopy.getFileSystemObjectFromPath(pathArr.join('/'));
-      } else {
-        // console.log(pathArr.join('/'));
-        dir = cwdCopy.getFileSystemObjectFromPath(pathArr.join('/'));
-      }
-    } else {
-      newFileName = path;
-      dir = cwdCopy;
-    }
-
-    if (!dir) {
-      return [`touch: '${path}': No such file or directory`];
-    } else if (!dir.isDirectory) {
-      return [`touch: '${path}': Not a directory`];
-    }
-
-    if (!(dir as Directory).getChild(newFileName)) {
-      (dir as Directory).addFileSystemObject(
-        directory ? new Directory(newFileName) : new File(newFileName)
-      );
-    } else if (directory) {
-      return [`mkdir: '${path}': File exists`];
-    }
-
-    prop.setFileSystem(fileSystemCopy);
-    prop.setCurrentWorkingDirectory(cwdCopy);
-    return [];
-  }
-
-  function executeMkdir(path: string): string[] {
-    return executeTouch(path, true);
-  }
-
-  // This is very sinful spaghetti code, sorry :(
-  interface CopyAndRmResult {
-    fileSystemCopy: Directory | null;
-    cwdCopy: Directory | null;
-    err: string[];
-  }
-
-  function executeRm(
-    path: string,
-    flags: string,
-    directory = false,
-    prevCwd?: Directory,
-    prevFileSystem?: Directory
-  ): CopyAndRmResult {
-    const result: CopyAndRmResult = {
-      fileSystemCopy: null,
-      cwdCopy: null,
-      err: [],
-    };
-
-    if (path === '.' || path === '..') {
-      result.err = ['rm: "." and ".." may not be removed'];
-      return result;
-    }
-    // Make a copy to avoid mutating the original (it's a state variable)
-    const fileSystemCopy = _.cloneDeep(
-      prevFileSystem ? prevFileSystem : prop.fileSystem
-    );
-    const cwdCopy = fileSystemCopy.getFileSystemObjectFromPath(
-      prevCwd ? prevCwd.path : prop.currentWorkingDirectory.path
-    ) as Directory;
-    let fileName = '';
-    let dir;
-
-    if (path.includes('/')) {
-      const pathArr = path.split('/');
-      fileName = pathArr.pop() || '';
-      if (path.startsWith('/')) {
-        dir = fileSystemCopy.getFileSystemObjectFromPath(pathArr.join('/'));
-      } else {
-        dir = cwdCopy.getFileSystemObjectFromPath(pathArr.join('/'));
-      }
-    } else {
-      fileName = path;
-      dir = cwdCopy;
-    }
-
-    const file = (dir as Directory)?.getChild(fileName);
-    if (!file) {
-      result.err = [`rm: '${path}': No such file or directory`];
-      return result;
-    }
-
-    const rmDirAndEmptyDirectory = directory && (file as Directory).isEmpty();
-    const removeDirectory = rmDirAndEmptyDirectory || flags.includes('r');
-
-    if (!file.isDirectory || removeDirectory) {
-      if (path === '/') {
-        result.err = [
-          "What are you doing? You can't delete the root directory!",
-        ];
-        // TODO: Figure out how to handle this
-        return result;
-      }
-      (dir as Directory).removeFileSystemObject(fileName);
-    } else {
-      if (directory && !(file as Directory).isEmpty()) {
-        result.err = [`rm: '${path}': Directory not empty`];
-        return result;
-      }
-      result.err = [`rm: '${path}': Is a directory`];
-      return result;
-    }
-
-    result.cwdCopy = cwdCopy;
-    result.fileSystemCopy = fileSystemCopy;
-    result.err = [];
-
-    prop.setFileSystem(fileSystemCopy);
-    prop.setCurrentWorkingDirectory(cwdCopy);
-    return result;
-  }
-
-  function executeRmdir(path: string, flags: string): string[] {
-    return executeRm(path, flags, true).err;
-  }
-
-  function executeCopy(
-    paths: string[],
-    destination: string,
-    flags: string
-  ): CopyAndRmResult {
-    const result: CopyAndRmResult = {
-      fileSystemCopy: null,
-      cwdCopy: null,
-      err: [],
-    };
-
-    // Make a copy to avoid mutating the original (it's a state variable)
-    const fileSystemCopy = _.cloneDeep(prop.fileSystem);
-    const cwdCopy = fileSystemCopy.getFileSystemObjectFromPath(
-      prop.currentWorkingDirectory.path
-    ) as Directory;
-    const newFiles: FileSystemObject[] = [];
-
-    // Determine whether to copy to a directory or rename the file
-    let destDir = destination.startsWith('/')
-      ? fileSystemCopy.getFileSystemObjectFromPath(destination)
-      : cwdCopy.getFileSystemObjectFromPath(destination);
-
-    let newFileName = '';
-    let rename = false;
-
-    if (paths.length > 1) {
-      if (!destDir) {
-        result.err = [`cp: '${destination}': No such file or directory`];
-        return result;
-      }
-      if (!destDir.isDirectory) {
-        result.err = [`cp: '${destination}': Not a directory`];
-        return result;
-      }
-    } else if (!destDir || !destDir.isDirectory) {
-      const tempDest = destination.split('/');
-      newFileName = tempDest.pop() || '';
-      destination = tempDest.join('/');
-
-      destDir = destination.startsWith('/')
-        ? fileSystemCopy.getFileSystemObjectFromPath(destination)
-        : cwdCopy.getFileSystemObjectFromPath(destination);
-
-      if (!destDir) {
-        result.err = [`cp: '${destination}': No such file or directory`];
-        return result;
-      }
-      if (!destDir.isDirectory) {
-        result.err = [`cp: '${destination}': Not a directory`];
-        return result;
-      }
-      rename = true;
-    }
-
-    for (const path of paths) {
-      const file = path.startsWith('/')
-        ? prop.fileSystem.getFileSystemObjectFromPath(path)
-        : prop.currentWorkingDirectory.getFileSystemObjectFromPath(path);
-
-      // If the path is absolute, find the directory to create the file in
-      // Otherwise, use the current working directory
-      newFileName = !rename ? path.split('/').pop() || '' : newFileName;
-
-      if (!file) {
-        result.err = [`cp: '${path}': No such file or directory`];
-        return result;
-      }
-      if (file.isDirectory && !flags.toLowerCase().includes('r')) {
-        result.err = [`cp: '${path}' is a directory (not copied)`];
-        return result;
-      }
-
-      const childDir = (destDir as Directory).getChild(newFileName);
-      const newFile = _.cloneDeep(file);
-      if (childDir && childDir.isDirectory) {
-        destDir = childDir;
-      } else {
-        newFile.name = newFileName;
-      }
-
-      newFiles.push(newFile);
-    }
-
-    for (const newFile of newFiles) {
-      (destDir as Directory).addFileSystemObject(newFile);
-    }
-
-    result.cwdCopy = cwdCopy;
-    result.fileSystemCopy = fileSystemCopy;
-
-    prop.setFileSystem(fileSystemCopy);
-    prop.setCurrentWorkingDirectory(cwdCopy);
-    return result;
-  }
-
-  function executeMove(paths: string[], destination: string): string[] {
-    // Make a copy to avoid mutating the original (it's a state variable)
-    let error: string[] = [];
-
-    // console.log(paths, destination, 'Executing copy part of move!');
-
-    let prevRes = executeCopy(paths, destination, '-r');
-    error = prevRes.err.map((err) => err.replace('cp', 'mv'));
-
-    // // If there are any errors in copy step, just return them and don't delete
-    if (error.length > 0) {
-      return error;
-    }
-
-    // Removing all the original files
-    for (const path of paths) {
-      prevRes = executeRm(
-        path,
-        '-r',
-        false,
-        // TODO: Change this logic here when refactoring code. This is incredibly cursed
-        prevRes.cwdCopy ? prevRes.cwdCopy : undefined,
-        prevRes.fileSystemCopy ? prevRes.fileSystemCopy : undefined
-      );
-      error = prevRes.err.map((err) => err.replace('rm', 'mv'));
-      if (error.length > 0) {
-        return error;
-      }
-    }
-
-    if (prevRes.fileSystemCopy) prop.setFileSystem(prevRes.fileSystemCopy);
-    if (prevRes.cwdCopy) prop.setCurrentWorkingDirectory(prevRes.cwdCopy);
-    return error;
-  }
-
-  function executeFind(path: string, flags: string): [string[], string[]] {
-    if (path || flags) {
-      return [[], ['find: Not implemented']];
-    }
-    return [[], []];
-  }
-
-  function isValid(usertyped: string) {
+  function processInput(usertyped: string) {
     const userInput = usertyped.trim().split(' ');
     const command = input.length > 1 ? userInput.slice(1, 2).join(' ') : '';
     const args = userInput.slice(2);
@@ -353,187 +30,27 @@ function Terminal(prop: {
       }
     }
     let path = args.slice(i).join(' ');
-    path = path === '' ? '.' : path;
-    let output: string[] = [];
-    let error: string[] = [];
+    path ||= '.';
 
-    switch (command) {
-      case 'whoami':
-        output = ['tux'];
-        break;
-      case 'pwd':
-        output = [prop.currentWorkingDirectory.path];
-        break;
-      case 'man':
-        switch (args[0]) {
-          case 'whoami':
-            output = [
-              `whoami – display effective user id
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'pwd':
-            output = [
-              `pwd – return working directory name
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'ls':
-            output = [
-              `ls – list directory contents.
-              OPTIONS: 
-                -a      Include directory entries whose names begin with a dot (‘.’).
-                -l      List files in the long format: file mode, number of links, owner name, group name, number of
-                bytes in the file, abbreviated month, day-of-month file was last
-                modified, hour file last modified, minute file last modified, and the
-                pathname.
-                (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'cd':
-            output = [
-              `cd — change the working directory
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'touch':
-            output = [
-              `touch – change file access and modification times
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'mkdir':
-            output = [
-              `mkdir – make directories
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'rmdir':
-            output = [
-              `rmdir – remove directories
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'cp':
-            output = [
-              `cp – copy files
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'mv':
-            output = [
-              `mv – move files
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'echo':
-            output = [
-              `echo – write arguments to the standard output
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'cat':
-            output = [
-              `cat – concatenate and print files
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'grep':
-            output = [
-              `grep - file pattern searcher. The grep utility searches any given input files, selecting lines that
-              match one or more patterns.  By default, a pattern matches an input line
-              if the regular expression (RE) in the pattern matches the input line
-              without its trailing newline.  An empty expression matches every line.
-              Each input line that matches at least one of the patterns is written to
-              the standard output.
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'find':
-            output = [
-              `find – walk a file hierarchy. The find utility recursively descends the directory tree for each path
-              listed, evaluating an expression in terms of each file in the tree.
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'chmod':
-            output = [
-              `chmod – change file modes or Access Control Lists.
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          case 'clear':
-            output = [
-              `clear - clear the terminal screen
-              (DESCRIPTION SIMPLIFIED FOR LAB)`,
-            ];
-            break;
-          default:
-            output = ['Command invalid for purposes of this learning lab'];
-        }
-        break;
-      case 'ls':
-        [output, error] = executeList(flags, path);
-        break;
-      case 'cd':
-        error = executeCd(path);
-        break;
-      case 'touch':
-        error = executeTouch(path);
-        break;
-      case 'mkdir':
-        error = executeMkdir(path);
-        break;
-      case 'rm':
-        error = executeRm(path, flags).err;
-        break;
-      case 'rmdir':
-        error = executeRmdir(path, flags);
-        break;
-      case 'cp': {
-        const cpArgs = path.split(' ');
-        if (cpArgs.length < 2) {
-          error = ['cp: invalid usage'];
-        } else {
-          const destination = cpArgs.pop();
-          // console.log(flags);
-          error = executeCopy(cpArgs, destination as string, flags).err;
-        }
-        break;
-      }
-      case 'mv': {
-        const mvArgs = path.split(' ');
-        if (mvArgs.length < 2) {
-          error = ['mv: invalid usage'];
-        } else {
-          const destination = mvArgs.pop();
-          error = executeMove(mvArgs, destination as string);
-        }
-        break;
-      }
-      case 'echo':
-        output = [path];
-        break;
-      case 'cat':
-        [output, error] = executeCat(path);
-        break;
-      case 'grep':
-        // console.log('User input grep');
-        break;
-      case 'find':
-        [output, error] = executeFind(path, flags);
-        break;
-      case 'chmod':
-        // console.log('User input chmod');
-        break;
-      case 'clear':
-        error = ['CLEAR'];
-        break;
-      default:
-        output = ['Invalid command. Try again.'];
-        break;
+    console.log('Command: ' + command, 'Flags: ' + flags, 'Path: ' + path);
+    const result: TerminalCommandResult = executeCommand(
+      _.cloneDeep(prop.fileSystem),
+      _.cloneDeep(prop.currentWorkingDirectory),
+      command,
+      path,
+      flags
+    );
+
+    const { modifiedCWD, modifiedFS, err, out } = result;
+
+    if (modifiedFS) {
+      prop.setFileSystem(modifiedFS);
     }
-    return { out: output, err: error };
+    if (modifiedCWD) {
+      prop.setCurrentWorkingDirectory(modifiedCWD);
+    }
+
+    return { err, out };
   }
 
   function handleChange(e: React.FormEvent<HTMLInputElement>) {
@@ -548,7 +65,7 @@ function Terminal(prop: {
 
   function handleSubmit(e: React.FormEvent) {
     e.preventDefault();
-    const result = isValid(input);
+    const result = processInput(input);
     const newHistory = [...inputHistory, input];
     if (result.err.length > 0 && result.err[0] === 'CLEAR') {
       setCommands([]);
