+++ src/components/shared/globalTypes.ts
@@ -158,6 +158,7 @@ export class Directory extends FileSystemObject {
     path: string
   ): Directory | string {
     currentWorkingDirectory.isCurrentDirectory = false;
+    console.log(currentWorkingDirectory);
     let newCwd: Directory | File | undefined;
 
     // If the path is absolute, we start from the root directory and find the new cwd
